# [SPARK-49596][SQL] Improve performance of `FormatString`