### What changes were proposed in this pull request?
The pr aims to improve performance of the expression `FormatString`.


### Why are the changes needed?
- In the expression `FormatString`, the class `java.lang.StringBuffer` used is `thread safe`, but its performance is not as good as `java.lang.StringBuilder`.
- In the scenario we are using, there is actually no need to use `java.lang.StringBuffer`.
- In class `java.util.Formatter`, it also defaults to using `java.lang.StringBuilder`, as follows:
  <img width="605" alt="image" src="https://github.com/user-attachments/assets/83f9e36a-3224-44f5-a905-164256589b54">

### Does this PR introduce _any_ user-facing change?
No.


### How was this patch tested?
Pass GA.
Manually check.


### Was this patch authored or co-authored using generative AI tooling?
No.